% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_calib.R
\name{calib}
\alias{calib}
\title{The function calib}
\usage{
calib(conc, OD)
}
\arguments{
\item{conc}{desc}

\item{OD}{desc}
}
\description{
SOme more explanation
}
\section{Warning}{

This function is not exported.
}

