% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc.R
\name{gc_content}
\alias{gc_content}
\title{Calculate GC content}
\usage{
gc_content(seq)
}
\arguments{
\item{seq}{sequence}
}
\value{
GC content
}
\description{
aad af adf adf a
}
\examples{
seq = "ATGCATGACAGATC"
gc(seq = seq)
}
