% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_get_conc.R
\name{get_conc}
\alias{get_conc}
\title{Get a new concenctration.}
\usage{
get_conc(conc, OD, new, ...)
}
\arguments{
\item{conc}{takes a numeric vector with concentrations.}

\item{OD}{takes the optical densities}

\item{new}{takes a vector with now ODs that should get a concentration.}

\item{...}{currently disregarded}
}
\value{
the predictions
}
\description{
It calculates a new concentration based on
a calibration curve. It uses internally the function \code{\link{calib}}.
}
\details{
Here you can give more details

Those details they can be multiple paragraphs.
}
\examples{
conc <- 1:10
OD <- 10:1
get_conc(conc, OD, 4.5)

\dontrun{
get_co("error")
}

}
\seealso{
\code{\link{calib}}.
}
