% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_build_fct.R
\name{myMAcomp}
\alias{myMAcomp}
\title{Row-wise summary stats}
\usage{
myMAcomp(myMA = myMA, group = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4), myfct = mean)
}
\arguments{
\item{myMA}{PARAM_DESCRIPTION, Default: myMA}

\item{group}{PARAM_DESCRIPTION, Default: c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4)}

\item{myfct}{PARAM_DESCRIPTION, Default: mean}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Computes row-wise summary statistics for any combination of
columns of a \code{matrix}. The first argument (\code{myMA})
specifies the input \code{matrix}; the second one (\code{group}) selects the columns
based on a grouping vector, and the third one the summary statistics (e.g. mean, sd, max).
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
   myMA <- matrix(rnorm(100000), 10000, 10, dimnames=list(1:10000, paste("C", 1:10, sep="")))
   myMAcomp(myMA=myMA, group=c(1,1,1,2,2,2,3,3,4,4), myfct=mean)[1:2,] 
 }
}
}
