% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmatch.R
\name{re_match}
\alias{re_match}
\alias{re_matches}
\title{Find regular expression matches}
\usage{
re_match(x, pat, perl = NULL, ...)

re_matches(x, pat, perl = NULL, ...)
}
\arguments{
\item{x}{Input character vector}

\item{pat}{Regular expression pattern/search string}

\item{perl}{Logical indicating whether to use perl-style regex.
If NULL (default) this checks for perl-style expressions.}

\item{...}{Other arguments passed to \code{\link{regexpr}} or
\code{\link{gregexpr}}}
}
\value{
A list containing character vector matches for each input
}
\description{
Return first or all regular expression matches
}
\examples{

## find first match
re_match(c("AbC", "dEf"), "[A-Z]")

## find all matches
re_matches(c("AbC", "dEf"), "[A-Z]")

}
