% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{subtract_numbers}
\alias{subtract_numbers}
\title{Subtract Two Numbers}
\usage{
subtract_numbers(a, b)
}
\arguments{
\item{a}{A numeric value representing the number to be subtracted from.}

\item{b}{A numeric value representing the number to subtract.}
}
\value{
A numeric value, which is the result of \code{a - b}.
}
\description{
This function subtracts the second numeric value from the first and returns the result.
}
\examples{
subtract_numbers(5, 2)  # Returns 3
}
