% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_build_fct.R
\name{talkToMe}
\alias{talkToMe}
\title{Talk to me function}
\usage{
talkToMe(talk = TRUE)
}
\arguments{
\item{talk}{Are you in the mood to talk? Defaults to TRUE.}
}
\description{
This function triggers response from your computer.
}
\examples{
talkToMe()
}
\keyword{talk}
